Moved VM struct to its own file, fixed clippy warnings
